backend/requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
asyncio
flask
openai>=1.0.0
pinecone
requests
flask-cors
dirtyjson
scikit-image
scipy
numpy
imagehash
Pillow>=9.0.0
opencv-python>=4.8.0
